Only draw one part of the tileset

+++ src/constants.rs
@@ -5,7 +5,7 @@ pub const SCREEN_WIDTH: f32 = 1280.0;
 pub const SCREEN_HEIGHT: f32 = 720.0;
 
 // Sprite dimensions
-pub const SPRITE_WIDTH: u32 = 128;
-pub const SPRITE_HEIGHT: u32 = 128;
+pub const SPRITE_WIDTH: f32 = 128.0;
+pub const SPRITE_HEIGHT: f32 = 128.0;
 
 pub const DRAW_BBOX: bool = false;
